integration with evaluation bundle
