+++ lib/evaluation/bundlers/index.ts
@@ -1,2 +1,3 @@
 export * from './evaluationBundler';
 export * from './presentationDefinitionEB';
+export * from './inputDescriptorsEB';
