+++ lib/evaluation/evaluationEngine.ts
@@ -11,7 +11,7 @@ export class Evaluator {
 export class EvaluationEngine {
   //TODO change this Object to PresentationSubmissionWrapper
   evaluate(evaluators: Evaluator[]) {
-    let evaluations: Evaluation<any, PresentationDefinition>[] = [];
+    let evaluations: Evaluation<any, EvaluationBundler<any, any>>[] = [];
 
     for (const evaluator of evaluators) {
       evaluations = evaluations.concat(evaluator.bundler.getEvaluations(evaluator.target.d, evaluator.target.p));
